Clean up scheduler client in em
