+++ components/spider-execution-manager/src/client/grpc/mod.rs
@@ -5,5 +5,5 @@ pub mod scheduler;
 pub mod storage;
 
 pub use liveness::GrpcLivenessClient;
-pub use scheduler::{GrpcSchedulerClient, next_task_request, scheduler_response_to_result};
+pub use scheduler::GrpcSchedulerClient;
 pub use storage::GrpcStorageClient;
